requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Database
SQLAlchemy>=2.0.0
sqlite3worker>=1.1.7  # Thread-safe SQLite

# Optimization
optuna>=3.5.0
joblib>=1.3.0  # For parallel execution

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0  # Coverage reporting
pytest-xdist>=3.3.0  # Parallel test execution

# Backtesting & Portfolio
vectorbt>=0.26.0  # Fast vectorized backtesting
# Alternative: backtrader>=1.9.78

# Statistical Analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# ML (Optional - for ML-based signals)
xgboost>=2.0.0
lightgbm>=4.1.0

# Technical Indicators
ta-lib>=0.4.28  # Note: Requires TA-Lib C library
# Alternative pure-Python: pandas-ta>=0.3.14

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0  # For Optuna visualizations

# Data Fetching (for crypto)
ccxt>=4.1.0  # Unified crypto exchange API
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0  # Environment variable management
pyyaml>=6.0  # Config file parsing
tqdm>=4.66.0  # Progress bars
loguru>=0.7.0  # Better logging

# Type Checking & Code Quality (Dev)
mypy>=1.7.0
black>=23.11.0  # Code formatting
flake8>=6.1.0  # Linting
isort>=5.12.0  # Import sorting

# Performance
numba>=0.58.0  # JIT compilation for hot loops
bottleneck>=1.3.7  # Fast array operations

# Development
ipython>=8.17.0
jupyter>=1.0.0
notebook>=7.0.0
